Headerの作成 Readerの削除
