+++ 音ゲー（仮）/Music.swift
@@ -8,18 +8,21 @@
 
 import Foundation
 
-/// bmsファイルから読み込まれた音楽情報をまとめたentity.readerが読み込んで完成する（読み込み前の扱いに注意）。
+/// bmsファイルから読み込まれた音楽情報をまとめたentity。ヘッダ情報とメイン情報を持つ。
 struct Music {
-    var laneNum = 7
-    var genre = ""                                      // ジャンル
-    var title = ""                                      // タイトル(正式名称。ファイル名は文字の制約があるためこっちを正式とする)
-    var artist = ""                                     // アーティスト
-    var videoID = ""                                    // YouTubeのvideoID
-    var playLevel = 0                                   // 難易度
-    var volWav = 100                                    // 音量を現段階のn%として出力するか(TODO: 未実装)
-    var BPMs: [(bpm: Double, startPos: Double)] = []    // 可変BPM情報
     
-    var musicName: MusicName {
-        return MusicName(rawValue: title)!
-    }                  // 曲名
+    var header: Header
+    
+    var laneNum:   Int    { return header.laneNum   }
+    var genre:     String { return header.genre     }   // ジャンル
+    var title:     String { return header.title     }   // タイトル(正式名称。ファイル名は文字の制約があるためこっちを正式とする)
+    var artist:    String { return header.artist    }   // アーティスト
+    var videoID:   String { return header.videoID   }   // YouTubeのvideoID
+    var playLevel: Int    { return header.playLevel }   // 難易度
+    var volWav:    Int    { return header.volWav    }   // 音量を現段階のn%として出力するか(TODO: 未実装)
+    var BPMs: [(bpm: Double, startPos: Double)]{        // 可変BPM情報(headerだけでなくmain情報にも記述されるのでget onlyにしない)
+        get { return header.BPMs }
+        set { header.BPMs = newValue }
+    }
+    var musicName: MusicName { return MusicName(rawValue: title)! }       // 曲名
 }
